requirements.txt
streamlit
networkx
google-generativeai
chromadb
pydantic
streamlit-agraph
scikit-learn
sentence-transformers
watchdog
pyvis
python-dotenv
pysqlite3-binary
